# yiyan_plugin_TravelHelper/.well-known/openapi.yaml
openapi: 3.0.1
info:
    title: 旅游助手
    description: 个性化的旅游助手，可以根据用户的提问为用户提供景点推荐和信息介绍，还可以根据用户的位置为用户提供路线推荐和出行方式推荐。
    version: "v1"
servers:
    - url: http://127.0.0.1:8081
paths:
    /get_spot:
        post:
            operationId: getSpot
            summary: 展示推荐的景点
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/getSpot"
            responses:
                "200":
                    description: 景点推荐完成
                    content: 
                        application/json:
                            schema:
                                $ref: "#/components/schemas/spotResponse"                
    /get_way:
        post:
            operationId: getWay
            summary: 展示推荐的出行方式
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/getWay"
            responses:
                "200":
                    description: 出行方式推荐完成
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/wayResponse"
    /get_wordbook:
        get:
            operationId: getWordbook
            summary: 展示单词列表
            responses:
                "200":
                    description: 列表展示完成
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/getWordbook"
    /generate_sentences:
        post:        
            operationId: generateSentences
            summary: 背单词，生成句子
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/generateSentences"
            responses:
                "200":
                    description: 句子生成成功
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/responseSentences"
    /add_word:
        post:
            operationId: addWord
            summary: 在单词本中添加一个单词
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/addWord"
            responses:
                "200":
                    description: 单词添加成功
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/messageResponse"
    /delete_word:
        delete:
            operationId: deleteWord
            summary: 从单词本中删除一个单词
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/deleteWord"
            responses:
                "200":
                    description: 单词删除成功
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/messageResponse"

components:
    schemas:
        getSpot:
            type: object
            required: [city]
            properties: 
                city:
                    type: string
                    description: 城市
        spotResponse:
            type: object
            required: [spot]
            properties:
                spot:
                    type: object
                    description: 景点信息
        getWay:
            type: object
            required: [cities]
            properties:
                cities:
                    type: array
                    description: 起点城市名和终点城市名
                    items:
                        type: string
        wayResponse:
            type: object
            required: [way]
            properties:
                way:
                    type: object
                    description: 出行方式
        getWordbook:
            type: object
            required: [wordbook]
            properties:
                wordbook:
                    type: array
                    items:
                        type: string
                    description: 单词本单词列表
        addWord:
            type: object
            required: [word]
            properties:
                word:
                    type: string
                    description: 需要添加到单词本中的一个单词
        deleteWord:
            type: object
            required: [word]
            properties:
                word:
                    type: string
                    description: 需要删除的单词
        generateSentences:
            type: object
            required: [word_number]
            properties:
                word_number:
                    type: integer
                    description: 几个单词
        responseSentences:
            type: object
            required: [words]
            properties:
                words:
                    type: array
                    items:
                        type: string
                        description: 随机抽取的英文单词
        messageResponse:
            type: object
            required: [message]
            properties:
                message:
                    type: string
                    description: 回复信息